madness I say, madness! (trying to do the grafana)

+++ src/metrics.js
@@ -128,7 +128,7 @@ function sendMetricToGrafana(metricName, metricValue, type, unit) {
                   [type]: {
                     dataPoints: [
                       {
-                        asInt: metricValue,
+                        asDouble: metricValue,
                         timeUnixNano: Date.now() * 1000000,
                       },
                     ],
